requirements.txt
python-evtx
python-dateutil